Add files via upload
homepage: there's a jumbotron behind reading challenge and book of the week  and below new authors there are space for  three new author's pic which will fade when we hover over it and some data is gonna displayso check that. only FANTASY is linked to other files so try that. and there is gonna be some pic {it's a grid system} behind every genre name. that best seller is horizontal side bar {forgot to add more products {will work when we add}}

last single product thing page: there's gonna be book cover at the left side.